+++ tests/lib/test_mastodon_helper.py
@@ -1,10 +1,12 @@
 from pyxavi.config import Config
 from janitor.lib.mastodon_helper import MastodonHelper
+from janitor.objects.mastodon_connection_params import MastodonConnectionParams
 from unittest.mock import patch, Mock
 from unittest import TestCase
 import pytest
 from mastodon import Mastodon
 import os
+from pyxavi.debugger import dd
 
 CONFIG = {
     "logger.name": "logger_test",
@@ -14,6 +16,19 @@ CONFIG = {
     "mastodon.api_base_url": "https://mastodont.cat",
     "mastodon.credentials.user.email": "bot+syscheck@my-fancy.site",
     "mastodon.credentials.user.password": "SuperSecureP4ss",
+}
+
+CONFIG_MASTODON_CONN_PARAMS = {
+    "instance_type": "mastodon",
+    "api_base_url": "https://mastodont.cat",
+    "credentials": {
+        "user_file": "user.secret",
+        "client_file": "client.secret",
+        "user": {
+            "email": "bot+syscheck@my-fancy.site",
+            "password": "SuperSecureP4ss",
+        }
+    }
 }
 
 
@@ -30,6 +45,7 @@ def patched_config_get(self, param: str, default=None) -> str:
     argvalues=[
         ("mastodon", MastodonHelper.TYPE_MASTODON, False),
         ("pleroma", MastodonHelper.TYPE_PLEROMA, False),
+        ("firefish", MastodonHelper.TYPE_FIREFISH, False),
         ("exception", None, RuntimeError),
     ],
 )
@@ -45,7 +61,7 @@ def test_message_type_valid_or_raise(value, expected_type, expected_exception):
 @patch.object(Config, "__init__", new=patched_config_init)
 @patch.object(Config, "get", new=patched_config_get)
 def test_get_instance_mastodon_user_credentials_exists():
-    CONFIG["mastodon.instance_type"] = "mastodon"
+    CONFIG_MASTODON_CONN_PARAMS["instance_type"] = "mastodon"
     mocked_path_exists = Mock()
     mocked_path_exists.return_value = True
     mocked_mastodon_init = Mock()
@@ -53,11 +69,16 @@ def test_get_instance_mastodon_user_credentials_exists():
     mocked_mastodon_init.__class__ = Mastodon
     with patch.object(os.path, "exists", new=mocked_path_exists):
         with patch.object(Mastodon, "__init__", new=mocked_mastodon_init):
-            instance = MastodonHelper.get_instance(config=Config())
+            conn_params = MastodonConnectionParams.from_dict(CONFIG_MASTODON_CONN_PARAMS)
+            dd(conn_params)
+            instance = MastodonHelper.get_instance(
+                config=Config(),
+                connection_params=conn_params
+            )
 
-    mocked_path_exists.assert_called_once_with(CONFIG["mastodon.credentials.user_file"])
+    mocked_path_exists.assert_called_once_with(CONFIG_MASTODON_CONN_PARAMS["credentials"]["user_file"])
     mocked_mastodon_init.assert_called_once_with(
-        access_token=CONFIG["mastodon.credentials.user_file"], feature_set="mainline"
+        access_token=CONFIG_MASTODON_CONN_PARAMS["credentials"]["user_file"], feature_set="mainline"
     )
     assert isinstance(instance, Mastodon)
 
@@ -65,7 +86,7 @@ def test_get_instance_mastodon_user_credentials_exists():
 @patch.object(Config, "__init__", new=patched_config_init)
 @patch.object(Config, "get", new=patched_config_get)
 def test_get_instance_pleroma_user_credentials_exists():
-    CONFIG["mastodon.instance_type"] = "pleroma"
+    CONFIG_MASTODON_CONN_PARAMS["instance_type"] = "pleroma"
     mocked_path_exists = Mock()
     mocked_path_exists.return_value = True
     mocked_mastodon_init = Mock()
@@ -73,11 +94,15 @@ def test_get_instance_pleroma_user_credentials_exists():
     mocked_mastodon_init.__class__ = Mastodon
     with patch.object(os.path, "exists", new=mocked_path_exists):
         with patch.object(Mastodon, "__init__", new=mocked_mastodon_init):
-            instance = MastodonHelper.get_instance(config=Config())
+            conn_params = MastodonConnectionParams.from_dict(CONFIG_MASTODON_CONN_PARAMS)
+            instance = MastodonHelper.get_instance(
+                config=Config(),
+                connection_params=conn_params
+            )
 
-    mocked_path_exists.assert_called_once_with(CONFIG["mastodon.credentials.user_file"])
+    mocked_path_exists.assert_called_once_with(CONFIG_MASTODON_CONN_PARAMS["credentials"]["user_file"])
     mocked_mastodon_init.assert_called_once_with(
-        access_token=CONFIG["mastodon.credentials.user_file"], feature_set="pleroma"
+        access_token=CONFIG_MASTODON_CONN_PARAMS["credentials"]["user_file"], feature_set="pleroma"
     )
     assert isinstance(instance, Mastodon)
 
@@ -85,7 +110,7 @@ def test_get_instance_pleroma_user_credentials_exists():
 @patch.object(Config, "__init__", new=patched_config_init)
 @patch.object(Config, "get", new=patched_config_get)
 def test_get_instance_mastodon_user_credentials_not_exists():
-    CONFIG["mastodon.instance_type"] = "mastodon"
+    CONFIG_MASTODON_CONN_PARAMS["instance_type"] = "mastodon"
     mocked_path_exists = Mock()
     mocked_path_exists.return_value = False
     mocked_mastodon_init = Mock()
@@ -95,18 +120,22 @@ def test_get_instance_mastodon_user_credentials_not_exists():
     with patch.object(os.path, "exists", new=mocked_path_exists):
         with patch.object(Mastodon, "__init__", new=mocked_mastodon_init):
             with patch.object(Mastodon, "log_in", new=mocked_mastodon_log_in):
-                instance = MastodonHelper.get_instance(config=Config())
+                conn_params = MastodonConnectionParams.from_dict(CONFIG_MASTODON_CONN_PARAMS)
+                instance = MastodonHelper.get_instance(
+                    config=Config(),
+                    connection_params=conn_params
+                )
 
-    mocked_path_exists.assert_called_once_with(CONFIG["mastodon.credentials.user_file"])
+    mocked_path_exists.assert_called_once_with(CONFIG_MASTODON_CONN_PARAMS["credentials"]["user_file"])
     mocked_mastodon_init.assert_called_once_with(
-        client_id=CONFIG["mastodon.credentials.client_file"],
-        api_base_url=CONFIG["mastodon.api_base_url"],
+        client_id=CONFIG_MASTODON_CONN_PARAMS["credentials"]["client_file"],
+        api_base_url=CONFIG_MASTODON_CONN_PARAMS["api_base_url"],
         feature_set="mainline"
     )
     mocked_mastodon_log_in.assert_called_once_with(
-        CONFIG["mastodon.credentials.user.email"],
-        CONFIG["mastodon.credentials.user.password"],
-        to_file=CONFIG["mastodon.credentials.user_file"]
+        CONFIG_MASTODON_CONN_PARAMS["credentials"]["user"]["email"],
+        CONFIG_MASTODON_CONN_PARAMS["credentials"]["user"]["password"],
+        to_file=CONFIG_MASTODON_CONN_PARAMS["credentials"]["user_file"]
     )
     assert isinstance(instance, Mastodon)
 
@@ -114,7 +143,7 @@ def test_get_instance_mastodon_user_credentials_not_exists():
 @patch.object(Config, "__init__", new=patched_config_init)
 @patch.object(Config, "get", new=patched_config_get)
 def test_get_instance_pleroma_user_credentials_not_exists():
-    CONFIG["mastodon.instance_type"] = "pleroma"
+    CONFIG_MASTODON_CONN_PARAMS["instance_type"] = "pleroma"
     mocked_path_exists = Mock()
     mocked_path_exists.return_value = False
     mocked_mastodon_init = Mock()
@@ -124,17 +153,21 @@ def test_get_instance_pleroma_user_credentials_not_exists():
     with patch.object(os.path, "exists", new=mocked_path_exists):
         with patch.object(Mastodon, "__init__", new=mocked_mastodon_init):
             with patch.object(Mastodon, "log_in", new=mocked_mastodon_log_in):
-                instance = MastodonHelper.get_instance(config=Config())
+                conn_params = MastodonConnectionParams.from_dict(CONFIG_MASTODON_CONN_PARAMS)
+                instance = MastodonHelper.get_instance(
+                    config=Config(),
+                    connection_params=conn_params
+                )
 
-    mocked_path_exists.assert_called_once_with(CONFIG["mastodon.credentials.user_file"])
+    mocked_path_exists.assert_called_once_with(CONFIG_MASTODON_CONN_PARAMS["credentials"]["user_file"])
     mocked_mastodon_init.assert_called_once_with(
-        client_id=CONFIG["mastodon.credentials.client_file"],
-        api_base_url=CONFIG["mastodon.api_base_url"],
+        client_id=CONFIG_MASTODON_CONN_PARAMS["credentials"]["client_file"],
+        api_base_url=CONFIG_MASTODON_CONN_PARAMS["api_base_url"],
         feature_set="pleroma"
     )
     mocked_mastodon_log_in.assert_called_once_with(
-        CONFIG["mastodon.credentials.user.email"],
-        CONFIG["mastodon.credentials.user.password"],
-        to_file=CONFIG["mastodon.credentials.user_file"]
+        CONFIG_MASTODON_CONN_PARAMS["credentials"]["user"]["email"],
+        CONFIG_MASTODON_CONN_PARAMS["credentials"]["user"]["password"],
+        to_file=CONFIG_MASTODON_CONN_PARAMS["credentials"]["user_file"]
     )
     assert isinstance(instance, Mastodon)