+++ CHANGELOG.md
@@ -1,15 +1,27 @@
 # Changelog
 
+## [v0.4.0](https://github.com/XaviArnaus/janitor/releases/tag/v0.4.0)
+
+### Added
+
+- Runner to monitor Git repositories and publish the identified new version from the Changelog
+- New `TODO.md` list ;-)
+
+### Changed
+
+- `MastodonHelper` now is able to receive connection params injected
+- Bumped pyxavi to 0.5.1
+
 ## [v0.3.0](https://github.com/XaviArnaus/janitor/releases/tag/v0.3.0)
 
 ### Added
 
 - Runner to update Directic's Dynamic DNS
-- Bumped pyxavi to 0.5.0
 
 ### Changed
 
 - Changelog iterated to adhere to [Common Changelog](https://common-changelog.org). Will start from this version on.
+- Bumped pyxavi to 0.5.0
 
 ## [v0.2.0](https://github.com/XaviArnaus/janitor/releases/tag/v0.2.0)